dom/html/test/test_focus_preventFocusRing.html

<!doctype html>
<title>Test for FocusOption.preventFocusRing</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<button>ABC</button>
<script>
  let button = document.querySelector("button");
  SpecialPowers.wrap(button).focus({ preventFocusRing: true });
  is(document.activeElement, button, "Button should be focused");
  ok(button.matches(":focus"), "Button should match :focus");
  ok(!button.matches(":focus-visible"), "Button should not match :focus-visible");
</script>
